fix(pptx): chart #,##0 number format rounds instead of truncating
ApplyNumFmt cast the value to (long) before formatting with N0, truncating
toward zero: a data label or axis tick of 1234.6 rendered as 1,234 while real
PowerPoint rounds to 1,235. Format the double directly with N0 so rounding
matches PowerPoint.

+++ src/officecli/Core/Chart/ChartSvgRenderer.cs
@@ -1988,7 +1988,7 @@ internal partial class ChartSvgRenderer
         if (useThousands)
             formatted = decimals > 0
                 ? val.ToString($"N{decimals}")
-                : ((long)val).ToString("N0");
+                : val.ToString("N0");  // round, don't truncate ((long) cast truncated 1234.6 -> 1234)
         else
             formatted = decimals > 0
                 ? val.ToString($"F{decimals}")
